readme: add known problems
The layout on the website was updated, but the tool is not updated
accordingly at the moment.

Grabbing of the title should be simple, but I don't see how I can deduce
an article number (if it is not set in the argument) aside of parsing
images URLs. Postponed for now, proposed a workaround.