Improve <blockquote/> parsing
The changes are motivated by quotes in https://what-if.xkcd.com/160/

The following is changed:

* Parse `<p/>` inside `<blockquote/>`.
* Produce a whitespace for a newline in the source HTML (but not after
  `<br/>`).
* Don't assume that `<blockquote/>` produces a single-line Markdown
  output (it may have multiple paragraphs).
* Produce double-asterisk/double-underscore marks for `<b/>` tag (it is
  the same as for `<strong/>`).

Many old articles have newlines in the source HTML and the Markdown
output for them now looks much better.

The main criteria here is informal: whether the resulting Markdown looks
better. In my opinion the updated script meets it.